chore: rename server/commands to server/slashcommands, move bot-helpers to server/lib
The Phase 1 target folder is renamed to `slashcommands/` to match the
existing app/slashcommands-* naming and make the folder's purpose obvious.
`bot-helpers/` is pulled out into `server/lib/bot-helpers/` since it isn't
a slash command — it got bundled into the Phase 1 commit by mistake.

Also updates MIGRATION_PLAN.md and the disposable Phase 0/1 migration
scripts to reflect the new target paths.
