+++ apps/meteor/scripts/migration/phase1-commands.tsv
@@ -1,20 +1,20 @@
 # Phase 1: Slash commands migration
 # Format: old-path<TAB>new-path (relative to apps/meteor/)
-app/slashcommand-asciiarts/server	server/commands/asciiarts
-app/slashcommands-archiveroom/server	server/commands/archiveroom
-app/slashcommands-ban/server	server/commands/ban
-app/slashcommands-create/server	server/commands/create
-app/slashcommands-help/server	server/commands/help
-app/slashcommands-hide/server	server/commands/hide
-app/slashcommands-invite/server	server/commands/invite
-app/slashcommands-inviteall/server	server/commands/inviteall
-app/slashcommands-join/server	server/commands/join
-app/slashcommands-kick/server	server/commands/kick
-app/slashcommands-leave/server	server/commands/leave
-app/slashcommands-me/server	server/commands/me
-app/slashcommands-msg/server	server/commands/msg
-app/slashcommands-mute/server	server/commands/mute
-app/slashcommands-status/server	server/commands/status
-app/slashcommands-topic/server	server/commands/topic
-app/slashcommands-unarchiveroom/server	server/commands/unarchiveroom
-app/bot-helpers/server	server/commands/bot-helpers
+app/slashcommand-asciiarts/server	server/slashcommands/asciiarts
+app/slashcommands-archiveroom/server	server/slashcommands/archiveroom
+app/slashcommands-ban/server	server/slashcommands/ban
+app/slashcommands-create/server	server/slashcommands/create
+app/slashcommands-help/server	server/slashcommands/help
+app/slashcommands-hide/server	server/slashcommands/hide
+app/slashcommands-invite/server	server/slashcommands/invite
+app/slashcommands-inviteall/server	server/slashcommands/inviteall
+app/slashcommands-join/server	server/slashcommands/join
+app/slashcommands-kick/server	server/slashcommands/kick
+app/slashcommands-leave/server	server/slashcommands/leave
+app/slashcommands-me/server	server/slashcommands/me
+app/slashcommands-msg/server	server/slashcommands/msg
+app/slashcommands-mute/server	server/slashcommands/mute
+app/slashcommands-status/server	server/slashcommands/status
+app/slashcommands-topic/server	server/slashcommands/topic
+app/slashcommands-unarchiveroom/server	server/slashcommands/unarchiveroom
+app/bot-helpers/server	server/lib/bot-helpers
