+++ apps/meteor/server/importPackages.ts
@@ -5,7 +5,7 @@ import '../app/apple/server';
 import '../app/assets/server';
 import '../app/authorization/server';
 import '../app/autotranslate/server';
-import './commands/bot-helpers';
+import './lib/bot-helpers';
 import '../app/channel-settings/server';
 import '../app/cloud/server';
 import '../app/crowd/server';
@@ -45,23 +45,23 @@ import '../app/oauth2-server-config/server';
 import '../app/push-notifications/server';
 import '../app/retention-policy/server';
 import '../app/slackbridge/server';
-import './commands/archiveroom';
-import './commands/asciiarts';
-import './commands/ban';
-import './commands/create';
-import './commands/help';
-import './commands/hide';
-import './commands/invite';
-import './commands/inviteall';
-import './commands/join';
-import './commands/kick';
-import './commands/leave';
-import './commands/me';
-import './commands/msg';
-import './commands/mute';
-import './commands/status';
-import './commands/topic';
-import './commands/unarchiveroom';
+import './slashcommands/archiveroom';
+import './slashcommands/asciiarts';
+import './slashcommands/ban';
+import './slashcommands/create';
+import './slashcommands/help';
+import './slashcommands/hide';
+import './slashcommands/invite';
+import './slashcommands/inviteall';
+import './slashcommands/join';
+import './slashcommands/kick';
+import './slashcommands/leave';
+import './slashcommands/me';
+import './slashcommands/msg';
+import './slashcommands/mute';
+import './slashcommands/status';
+import './slashcommands/topic';
+import './slashcommands/unarchiveroom';
 import '../app/smarsh-connector/server';
 import '../app/theme/server';
 import '../app/threads/server';
