+++ apps/meteor/server/slashcommands/asciiarts/gimme.ts
