+++ apps/meteor/server/slashcommands/asciiarts/lenny.ts
