+++ apps/meteor/server/slashcommands/asciiarts/shrug.ts
