+++ apps/meteor/server/slashcommands/ban/ban.ts
