+++ apps/meteor/server/slashcommands/ban/index.ts
