+++ apps/meteor/server/slashcommands/help/index.ts
