+++ apps/meteor/server/slashcommands/help/server.ts
