+++ apps/meteor/server/slashcommands/hide/hide.ts
