+++ apps/meteor/server/slashcommands/hide/index.ts
