+++ apps/meteor/server/slashcommands/invite/index.ts
