+++ apps/meteor/server/slashcommands/inviteall/index.ts
