+++ apps/meteor/server/slashcommands/inviteall/server.ts
