+++ apps/meteor/server/slashcommands/join/index.ts
