+++ apps/meteor/server/slashcommands/join/server.ts
