+++ apps/meteor/server/slashcommands/kick/index.ts
