+++ apps/meteor/server/slashcommands/kick/server.ts
