+++ apps/meteor/server/slashcommands/leave/index.ts
