+++ apps/meteor/server/slashcommands/leave/leave.ts
