+++ apps/meteor/server/slashcommands/mute/index.ts
