+++ apps/meteor/server/slashcommands/mute/mute.ts
