+++ apps/meteor/server/slashcommands/status/status.ts
