+++ apps/meteor/server/slashcommands/unarchiveroom/index.ts
